@@ -6,11 +6,10 @@
     "dotenv": "^16.0.0",
     "firebase": "^9.6.9",
     "googleapis": "^99.0.0",
-    "jest": "^27.5.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-firebase-hooks": "^5.0.3",
-    "react-router-dom": "^6.2.2"
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-firebase-hooks": "^5.1.1",
+     "react-router-dom": "6.4.3"
   },
   "scripts": {
     "lint": "eslint .",
@@ -44,15 +43,16 @@
   },
   "devDependencies": {
     "@playwright/test": "1.47.0",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
     "browserstack-node-sdk": "^1.34.49",
     "danger": "^11.0.2",
     "eslint": "^8.11.0",
     "eslint-config-react-app": "^7.0.0",
     "eslint-plugin-playwright": "^0.8.0",
     "husky": "^4.3.7",
+    "jest": "^29.7.0",
     "msw": "^0.39.2",
     "react-scripts": "^5.0.1",
     "rimraf": "^3.0.2"
