@@ -28,38 +28,41 @@ describe('Login', () => {
     expect(screen.getByText(/remember me/i)).toBeInTheDocument();
   });
 
-  test('disables Sign in button if email or password is empty', () => {
+  test('disables Sign in button if email or password is empty', async () => {
+    const user = userEvent.setup();
     renderWithAuth();
 
     expect(screen.getByRole('button', { name: /sign in/i })).toBeDisabled();
 
-    userEvent.type(screen.getByLabelText('Username'), 'test@test.com' );
+    await user.type(screen.getByLabelText('Username'), 'test@test.com' );
     expect(screen.getByRole('button')).toBeDisabled();
 
-    userEvent.type(screen.getByLabelText('Password'), 'pass123' );
+    await user.type(screen.getByLabelText('Password'), 'pass123' );
     expect(screen.getByRole('button')).not.toBeDisabled();
   });
 
-  test('calls login with correct email, password, and rememberMe', () => {
+  test('calls login with correct email, password, and rememberMe', async () => {
+    const user = userEvent.setup();
     const loginMock = jest.fn();
     renderWithAuth({ login: loginMock });
 
-    userEvent.type(screen.getByLabelText('Username'), 'test@test.com' );
-    userEvent.type(screen.getByLabelText('Password'), 'pass123' );
-    userEvent.click(screen.getByRole('button', { name: /sign in/i }));
+    await user.type(screen.getByLabelText('Username'), 'test@test.com' );
+    await user.type(screen.getByLabelText('Password'), 'pass123' );
+    await user.click(screen.getByRole('button', { name: /sign in/i }));
 
     expect(loginMock).toHaveBeenCalledWith('test@test.com', 'pass123', true);
   });
 
-  test('toggles rememberMe when switch is clicked', () => {
+  test('toggles rememberMe when switch is clicked', async () => {
+    const user = userEvent.setup();
     const loginMock = jest.fn();
     renderWithAuth({ login: loginMock });
 
-    userEvent.type(screen.getByLabelText('Username'), 'test@test.com');
-    userEvent.type(screen.getByLabelText('Password'), 'pass123' );
+    await user.type(screen.getByLabelText('Username'), 'test@test.com');
+    await user.type(screen.getByLabelText('Password'), 'pass123' );
 
-    userEvent.click(screen.getByLabelText(/remember me/i));
-    userEvent.click(screen.getByRole('button', { name: /sign in/i }));
+    await user.click(screen.getByLabelText(/remember me/i));
+    await user.click(screen.getByRole('button', { name: /sign in/i }));
 
     expect(loginMock).toHaveBeenCalledWith('test@test.com', 'pass123', false);
   });
