@@ -29,14 +29,15 @@ describe('InstrumentForm component', () => {
     expect(screen.getByRole('textbox', { name: 'Description' })).toHaveValue('cool one');
   });
 
-  test('calls submitInstrument when form is submitted with valid inputs', () => {
+  test('calls submitInstrument when form is submitted with valid inputs', async () => {
+    const user = userEvent.setup();
     render(<InstrumentForm submitInstrument={submitInstrumentMock} closeModal={closeModalMock} />);
 
-    userEvent.type(screen.getByRole('textbox', { name: 'Name' }), 'Test gear name' );
-    userEvent.selectOptions(screen.getByLabelText('Type'), 'guitar' );
-    userEvent.type(screen.getByRole('textbox', { name: 'Description' }), 'Test description');
+    await user.type(screen.getByRole('textbox', { name: 'Name' }), 'Test gear name' );
+    await user.selectOptions(screen.getByLabelText('Type'), 'guitar' );
+    await user.type(screen.getByRole('textbox', { name: 'Description' }), 'Test description');
 
-    userEvent.click(screen.getByRole('button', { name: 'Add' }));
+    await user.click(screen.getByRole('button', { name: 'Add' }));
     expect(submitInstrumentMock).toHaveBeenCalledWith({
       name: 'Test gear name',
       type: 'guitar',
@@ -45,10 +46,11 @@ describe('InstrumentForm component', () => {
     });
   });
 
-  test('calls modal closed when cancel button is clicked', () => {
+  test('calls modal closed when cancel button is clicked', async () => {
+    const user = userEvent.setup();
     render(<InstrumentForm submitInstrument={submitInstrumentMock} closeModal={closeModalMock} />);
 
-    userEvent.click(screen.getByRole('button', { name: /cancel/i }));
+    await user.click(screen.getByRole('button', { name: /cancel/i }));
 
     expect(closeModalMock).toHaveBeenCalled();
   });
