@@ -25,25 +25,29 @@ describe('InstrumentListActions component', () => {
     expect(screen.getByRole('button', { name: 'Add new instrument' })).toBeInTheDocument();
   });
 
-  test('clicking filter button shows filter dropdown', () => {
-    userEvent.click(screen.getByRole('button', { name: 'Toggle filter menu' }));
+  test('clicking filter button shows filter dropdown', async () => {
+    const user = userEvent.setup();
+    await user.click(screen.getByRole('button', { name: 'Toggle filter menu' }));
     expect(screen.getByRole('checkbox', { name: 'guitar filter' })).toBeInTheDocument();
   });
 
-  test('clicking reset filters button calls resetFilter function', () => {
-    userEvent.click(screen.getByRole('button', { name: 'Toggle filter menu' }));
-    userEvent.click(screen.getByRole('button', { name: 'Reset Filters' }));
+  test('clicking reset filters button calls resetFilter function', async () => {
+    const user = userEvent.setup();
+    await user.click(screen.getByRole('button', { name: 'Toggle filter menu' }));
+    await user.click(screen.getByRole('button', { name: 'Reset Filters' }));
     expect(resetFilter).toHaveBeenCalledTimes(1);
   });
 
-  test('clicking add instrument button calls addInstrument function', () => {
-    userEvent.click(screen.getByRole('button', { name: 'Add new instrument' }));
+  test('clicking add instrument button calls addInstrument function', async () => {
+    const user = userEvent.setup();
+    await user.click(screen.getByRole('button', { name: 'Add new instrument' }));
     expect(addInstrument).toHaveBeenCalledTimes(1);
   });
 
-  test('clicking a filter switch calls updateFilter function', () => {
-    userEvent.click(screen.getByRole('button', { name: 'Toggle filter menu' }));
-    userEvent.click(screen.getByRole('checkbox', { name: 'guitar filter' }));
+  test('clicking a filter switch calls updateFilter function', async  () => {
+    const user = userEvent.setup();
+    await user.click(screen.getByRole('button', { name: 'Toggle filter menu' }));
+    await user.click(screen.getByRole('checkbox', { name: 'guitar filter' }));
 
     expect(updateFilter).toHaveBeenCalledTimes(1);
     expect(updateFilter).toHaveBeenCalledWith('guitar');
