@@ -2,7 +2,7 @@
 // allows you to do things like:
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
-import '@testing-library/jest-dom/extend-expect';
+import '@testing-library/jest-dom';
 import { server } from './mocks/server.js';
 import { configure } from '@testing-library/dom';
 require('dotenv').config(({ path: '.env.local' }));
